+++ packages/rotatable/CHANGELOG.md
@@ -4,4 +4,7 @@
 
 ### Minor Changes
 
-- [#106](https://github.com/willnguyen1312/headlessplayback/pull/106) [`a0d7433`](https://github.com/willnguyen1312/headlessplayback/commit/a0d7433ee819623c84b50a488af712f5d04ce2b4) Thanks [@willnguyen1312](https://github.com/willnguyen1312)! - Release rotatable plugin 💞
+- [#106](https://github.com/willnguyen1312/headlessplayback/pull/106)
+  [`a0d7433`](https://github.com/willnguyen1312/headlessplayback/commit/a0d7433ee819623c84b50a488af712f5d04ce2b4)
+  Thanks [@willnguyen1312](https://github.com/willnguyen1312)! - Release
+  rotatable plugin 💞